Move filterModal components to separate folder
